requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
Flask==3.0.3
Flask-Cors==4.0.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
openai==1.30.5
pydantic==2.7.2
pydantic_core==2.18.3
python-dotenv==1.0.1
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.1
Werkzeug==3.0.3
